+++ src/core/Odin.Core.Storage/SQLite/IdentityDatabase/TableCirclememberCRUD.cs
@@ -45,6 +45,35 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
     public class TableCircleMemberCRUD : TableBase
     {
         private bool _disposed = false;
+        private SqliteCommand _insertCommand = null;
+        private static Object _insertLock = new Object();
+        private SqliteParameter _insertParam1 = null;
+        private SqliteParameter _insertParam2 = null;
+        private SqliteParameter _insertParam3 = null;
+        private SqliteCommand _updateCommand = null;
+        private static Object _updateLock = new Object();
+        private SqliteParameter _updateParam1 = null;
+        private SqliteParameter _updateParam2 = null;
+        private SqliteParameter _updateParam3 = null;
+        private SqliteCommand _upsertCommand = null;
+        private static Object _upsertLock = new Object();
+        private SqliteParameter _upsertParam1 = null;
+        private SqliteParameter _upsertParam2 = null;
+        private SqliteParameter _upsertParam3 = null;
+        private SqliteCommand _delete0Command = null;
+        private static Object _delete0Lock = new Object();
+        private SqliteParameter _delete0Param1 = null;
+        private SqliteParameter _delete0Param2 = null;
+        private SqliteCommand _get0Command = null;
+        private static Object _get0Lock = new Object();
+        private SqliteParameter _get0Param1 = null;
+        private SqliteParameter _get0Param2 = null;
+        private SqliteCommand _get1Command = null;
+        private static Object _get1Lock = new Object();
+        private SqliteParameter _get1Param1 = null;
+        private SqliteCommand _get2Command = null;
+        private static Object _get2Lock = new Object();
+        private SqliteParameter _get2Param1 = null;
         private readonly CacheHelper _cache;
 
         public TableCircleMemberCRUD(IdentityDatabase db, CacheHelper cache) : base(db)
@@ -59,20 +88,33 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
 
         public override void Dispose()
         {
+            _insertCommand?.Dispose();
+            _insertCommand = null;
+            _updateCommand?.Dispose();
+            _updateCommand = null;
+            _upsertCommand?.Dispose();
+            _upsertCommand = null;
+            _delete0Command?.Dispose();
+            _delete0Command = null;
+            _get0Command?.Dispose();
+            _get0Command = null;
+            _get1Command?.Dispose();
+            _get1Command = null;
+            _get2Command?.Dispose();
+            _get2Command = null;
             _disposed = true;
-            GC.SuppressFinalize(this);
         }
 
-        public sealed override void EnsureTableExists(DatabaseConnection conn, bool dropExisting = false)
+        public sealed override void EnsureTableExists(bool dropExisting = false)
         {
-                using (var cmd = _database.CreateCommand())
+            using (var cmd = _database.CreateCommand())
+            {
+                if (dropExisting)
                 {
-                    if (dropExisting)
-                    {
-                       cmd.CommandText = "DROP TABLE IF EXISTS circleMember;";
-                       conn.ExecuteNonQuery(cmd);
-                    }
-                    cmd.CommandText =
+                    cmd.CommandText = "DROP TABLE IF EXISTS circleMember;";
+                    _database.ExecuteNonQuery(cmd);
+                }
+                cmd.CommandText =
                     "CREATE TABLE IF NOT EXISTS circleMember("
                      +"circleId BLOB NOT NULL, "
                      +"memberId BLOB NOT NULL, "
@@ -80,100 +122,106 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
                      +", PRIMARY KEY (circleId,memberId)"
                      +");"
                      ;
-                    conn.ExecuteNonQuery(cmd);
+                _database.ExecuteNonQuery(cmd);
+                _database.Commit();
             }
         }
 
-        public virtual int Insert(DatabaseConnection conn, CircleMemberRecord item)
+        public virtual int Insert(CircleMemberRecord item)
         {
-                using (var _insertCommand = _database.CreateCommand())
+            lock (_insertLock)
+            {
+                if (_insertCommand == null)
                 {
+                    _insertCommand = _database.CreateCommand();
                     _insertCommand.CommandText = "INSERT INTO circleMember (circleId,memberId,data) " +
                                                  "VALUES ($circleId,$memberId,$data)";
-                    var _insertParam1 = _insertCommand.CreateParameter();
-                    _insertParam1.ParameterName = "$circleId";
+                    _insertParam1 = _insertCommand.CreateParameter();
                     _insertCommand.Parameters.Add(_insertParam1);
-                    var _insertParam2 = _insertCommand.CreateParameter();
-                    _insertParam2.ParameterName = "$memberId";
+                    _insertParam1.ParameterName = "$circleId";
+                    _insertParam2 = _insertCommand.CreateParameter();
                     _insertCommand.Parameters.Add(_insertParam2);
-                    var _insertParam3 = _insertCommand.CreateParameter();
-                    _insertParam3.ParameterName = "$data";
+                    _insertParam2.ParameterName = "$memberId";
+                    _insertParam3 = _insertCommand.CreateParameter();
                     _insertCommand.Parameters.Add(_insertParam3);
+                    _insertParam3.ParameterName = "$data";
+                    _insertCommand.Prepare();
+                }
                 _insertParam1.Value = item.circleId.ToByteArray();
                 _insertParam2.Value = item.memberId.ToByteArray();
                 _insertParam3.Value = item.data ?? (object)DBNull.Value;
-                var count = conn.ExecuteNonQuery(_insertCommand);
+                var count = _database.ExecuteNonQuery(_insertCommand);
                 if (count > 0)
                  {
                     _cache.AddOrUpdate("TableCircleMemberCRUD", item.circleId.ToString()+item.memberId.ToString(), item);
                  }
                 return count;
-                } // Using
+            } // Lock
         }
 
-        public virtual int Upsert(DatabaseConnection conn, CircleMemberRecord item)
+        public virtual int Upsert(CircleMemberRecord item)
         {
-                using (var _upsertCommand = _database.CreateCommand())
+            lock (_upsertLock)
+            {
+                if (_upsertCommand == null)
                 {
+                    _upsertCommand = _database.CreateCommand();
                     _upsertCommand.CommandText = "INSERT INTO circleMember (circleId,memberId,data) " +
                                                  "VALUES ($circleId,$memberId,$data)"+
                                                  "ON CONFLICT (circleId,memberId) DO UPDATE "+
                                                  "SET data = $data "+
                                                  ";";
-                    var _upsertParam1 = _upsertCommand.CreateParameter();
-                    _upsertParam1.ParameterName = "$circleId";
+                    _upsertParam1 = _upsertCommand.CreateParameter();
                     _upsertCommand.Parameters.Add(_upsertParam1);
-                    var _upsertParam2 = _upsertCommand.CreateParameter();
-                    _upsertParam2.ParameterName = "$memberId";
+                    _upsertParam1.ParameterName = "$circleId";
+                    _upsertParam2 = _upsertCommand.CreateParameter();
                     _upsertCommand.Parameters.Add(_upsertParam2);
-                    var _upsertParam3 = _upsertCommand.CreateParameter();
-                    _upsertParam3.ParameterName = "$data";
+                    _upsertParam2.ParameterName = "$memberId";
+                    _upsertParam3 = _upsertCommand.CreateParameter();
                     _upsertCommand.Parameters.Add(_upsertParam3);
+                    _upsertParam3.ParameterName = "$data";
+                    _upsertCommand.Prepare();
+                }
                 _upsertParam1.Value = item.circleId.ToByteArray();
                 _upsertParam2.Value = item.memberId.ToByteArray();
                 _upsertParam3.Value = item.data ?? (object)DBNull.Value;
-                var count = conn.ExecuteNonQuery(_upsertCommand);
+                var count = _database.ExecuteNonQuery(_upsertCommand);
                 if (count > 0)
                     _cache.AddOrUpdate("TableCircleMemberCRUD", item.circleId.ToString()+item.memberId.ToString(), item);
                 return count;
-                } // Using
+            } // Lock
         }
-        public virtual int Update(DatabaseConnection conn, CircleMemberRecord item)
+        public virtual int Update(CircleMemberRecord item)
         {
-                using (var _updateCommand = _database.CreateCommand())
+            lock (_updateLock)
+            {
+                if (_updateCommand == null)
                 {
+                    _updateCommand = _database.CreateCommand();
                     _updateCommand.CommandText = "UPDATE circleMember " +
                                                  "SET data = $data "+
                                                  "WHERE (circleId = $circleId,memberId = $memberId)";
-                    var _updateParam1 = _updateCommand.CreateParameter();
-                    _updateParam1.ParameterName = "$circleId";
+                    _updateParam1 = _updateCommand.CreateParameter();
                     _updateCommand.Parameters.Add(_updateParam1);
-                    var _updateParam2 = _updateCommand.CreateParameter();
-                    _updateParam2.ParameterName = "$memberId";
+                    _updateParam1.ParameterName = "$circleId";
+                    _updateParam2 = _updateCommand.CreateParameter();
                     _updateCommand.Parameters.Add(_updateParam2);
-                    var _updateParam3 = _updateCommand.CreateParameter();
-                    _updateParam3.ParameterName = "$data";
+                    _updateParam2.ParameterName = "$memberId";
+                    _updateParam3 = _updateCommand.CreateParameter();
                     _updateCommand.Parameters.Add(_updateParam3);
+                    _updateParam3.ParameterName = "$data";
+                    _updateCommand.Prepare();
+                }
                 _updateParam1.Value = item.circleId.ToByteArray();
                 _updateParam2.Value = item.memberId.ToByteArray();
                 _updateParam3.Value = item.data ?? (object)DBNull.Value;
-                var count = conn.ExecuteNonQuery(_updateCommand);
+                var count = _database.ExecuteNonQuery(_updateCommand);
                 if (count > 0)
                 {
                     _cache.AddOrUpdate("TableCircleMemberCRUD", item.circleId.ToString()+item.memberId.ToString(), item);
                 }
                 return count;
-                } // Using
-        }
-
-        public virtual int GetCount(DatabaseConnection conn)
-        {
-                using (var _getCountCommand = _database.CreateCommand())
-                {
-                    _getCountCommand.CommandText = "PRAGMA read_uncommitted = 1; SELECT COUNT(*) FROM circleMember; PRAGMA read_uncommitted = 0;";
-                    var count = conn.ExecuteNonQuery(_getCountCommand);
-                    return count;
-                }
+            } // Lock
         }
 
         // SELECT circleId,memberId,data
@@ -222,26 +270,30 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
             return item;
        }
 
-        public int Delete(DatabaseConnection conn, Guid circleId,Guid memberId)
+        public int Delete(Guid circleId,Guid memberId)
         {
-                using (var _delete0Command = _database.CreateCommand())
+            lock (_delete0Lock)
+            {
+                if (_delete0Command == null)
                 {
+                    _delete0Command = _database.CreateCommand();
                     _delete0Command.CommandText = "DELETE FROM circleMember " +
                                                  "WHERE circleId = $circleId AND memberId = $memberId";
-                    var _delete0Param1 = _delete0Command.CreateParameter();
-                    _delete0Param1.ParameterName = "$circleId";
+                    _delete0Param1 = _delete0Command.CreateParameter();
                     _delete0Command.Parameters.Add(_delete0Param1);
-                    var _delete0Param2 = _delete0Command.CreateParameter();
-                    _delete0Param2.ParameterName = "$memberId";
+                    _delete0Param1.ParameterName = "$circleId";
+                    _delete0Param2 = _delete0Command.CreateParameter();
                     _delete0Command.Parameters.Add(_delete0Param2);
-
+                    _delete0Param2.ParameterName = "$memberId";
+                    _delete0Command.Prepare();
+                }
                 _delete0Param1.Value = circleId.ToByteArray();
                 _delete0Param2.Value = memberId.ToByteArray();
-                var count = conn.ExecuteNonQuery(_delete0Command);
+                var count = _database.ExecuteNonQuery(_delete0Command);
                 if (count > 0)
                     _cache.Remove("TableCircleMemberCRUD", circleId.ToString()+memberId.ToString());
                 return count;
-                } // Using
+            } // Lock
         }
 
         public CircleMemberRecord ReadRecordFromReader0(SqliteDataReader rdr, Guid circleId,Guid memberId)
@@ -271,27 +323,29 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
             return item;
        }
 
-        public CircleMemberRecord Get(DatabaseConnection conn, Guid circleId,Guid memberId)
+        public CircleMemberRecord Get(Guid circleId,Guid memberId)
         {
             var (hit, cacheObject) = _cache.Get("TableCircleMemberCRUD", circleId.ToString()+memberId.ToString());
             if (hit)
                 return (CircleMemberRecord)cacheObject;
-                using (var _get0Command = _database.CreateCommand())
+            lock (_get0Lock)
+            {
+                if (_get0Command == null)
                 {
+                    _get0Command = _database.CreateCommand();
                     _get0Command.CommandText = "SELECT data FROM circleMember " +
                                                  "WHERE circleId = $circleId AND memberId = $memberId LIMIT 1;";
-                    var _get0Param1 = _get0Command.CreateParameter();
-                    _get0Param1.ParameterName = "$circleId";
+                    _get0Param1 = _get0Command.CreateParameter();
                     _get0Command.Parameters.Add(_get0Param1);
-                    var _get0Param2 = _get0Command.CreateParameter();
-                    _get0Param2.ParameterName = "$memberId";
+                    _get0Param1.ParameterName = "$circleId";
+                    _get0Param2 = _get0Command.CreateParameter();
                     _get0Command.Parameters.Add(_get0Param2);
-
+                    _get0Param2.ParameterName = "$memberId";
+                    _get0Command.Prepare();
+                }
                 _get0Param1.Value = circleId.ToByteArray();
                 _get0Param2.Value = memberId.ToByteArray();
-                    lock (conn._lock)
-                    {
-                using (SqliteDataReader rdr = conn.ExecuteReader(_get0Command, System.Data.CommandBehavior.SingleRow))
+                using (SqliteDataReader rdr = _database.ExecuteReader(_get0Command, System.Data.CommandBehavior.SingleRow))
                 {
                     if (!rdr.Read())
                     {
@@ -303,7 +357,6 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
                     return r;
                 } // using
             } // lock
-            } // using
         }
 
         public CircleMemberRecord ReadRecordFromReader1(SqliteDataReader rdr, Guid circleId)
@@ -342,20 +395,22 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
             return item;
        }
 
-        public List<CircleMemberRecord> GetCircleMembers(DatabaseConnection conn, Guid circleId)
+        public List<CircleMemberRecord> GetCircleMembers(Guid circleId)
         {
-                using (var _get1Command = _database.CreateCommand())
+            lock (_get1Lock)
+            {
+                if (_get1Command == null)
                 {
+                    _get1Command = _database.CreateCommand();
                     _get1Command.CommandText = "SELECT memberId,data FROM circleMember " +
                                                  "WHERE circleId = $circleId;";
-                    var _get1Param1 = _get1Command.CreateParameter();
-                    _get1Param1.ParameterName = "$circleId";
+                    _get1Param1 = _get1Command.CreateParameter();
                     _get1Command.Parameters.Add(_get1Param1);
-
+                    _get1Param1.ParameterName = "$circleId";
+                    _get1Command.Prepare();
+                }
                 _get1Param1.Value = circleId.ToByteArray();
-                    lock (conn._lock)
-                    {
-                using (SqliteDataReader rdr = conn.ExecuteReader(_get1Command, System.Data.CommandBehavior.Default))
+                using (SqliteDataReader rdr = _database.ExecuteReader(_get1Command, System.Data.CommandBehavior.Default))
                 {
                     if (!rdr.Read())
                     {
@@ -372,7 +427,6 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
                     return result;
                 } // using
             } // lock
-            } // using
         }
 
         public CircleMemberRecord ReadRecordFromReader2(SqliteDataReader rdr, Guid memberId)
@@ -411,20 +465,22 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
             return item;
        }
 
-        public List<CircleMemberRecord> GetMemberCirclesAndData(DatabaseConnection conn, Guid memberId)
+        public List<CircleMemberRecord> GetMemberCirclesAndData(Guid memberId)
         {
-                using (var _get2Command = _database.CreateCommand())
+            lock (_get2Lock)
+            {
+                if (_get2Command == null)
                 {
+                    _get2Command = _database.CreateCommand();
                     _get2Command.CommandText = "SELECT circleId,data FROM circleMember " +
                                                  "WHERE memberId = $memberId;";
-                    var _get2Param1 = _get2Command.CreateParameter();
-                    _get2Param1.ParameterName = "$memberId";
+                    _get2Param1 = _get2Command.CreateParameter();
                     _get2Command.Parameters.Add(_get2Param1);
-
+                    _get2Param1.ParameterName = "$memberId";
+                    _get2Command.Prepare();
+                }
                 _get2Param1.Value = memberId.ToByteArray();
-                    lock (conn._lock)
-                    {
-                using (SqliteDataReader rdr = conn.ExecuteReader(_get2Command, System.Data.CommandBehavior.Default))
+                using (SqliteDataReader rdr = _database.ExecuteReader(_get2Command, System.Data.CommandBehavior.Default))
                 {
                     if (!rdr.Read())
                     {
@@ -441,7 +497,6 @@ namespace Odin.Core.Storage.SQLite.IdentityDatabase
                     return result;
                 } // using
             } // lock
-            } // using
         }
 
     }
